+++ src/index.ts
@@ -1,3 +1,6 @@
 export * from './types'
 export * from './ElectronDownloadManager'
 export * from './ElectronDownloadManagerMock'
+export { getFilenameFromMime } from "./utils";
+export { generateRandomId } from "./utils";
+export { truncateUrl } from "./utils";
